Assignment

+++ src/main/java/com/driver/OrderService.java
@@ -104,6 +104,7 @@ public class OrderService {
     }
 
     public void deletePartner(String partnerId){
+        if(!orderRepository.partnerDb.containsKey(partnerId)) return;
         orderRepository.partnerDb.remove(partnerId);
         for(String orderId : orderRepository.orderPartnerPairDb.keySet()){
             if(orderRepository.orderPartnerPairDb.get(orderId).equals(partnerId)){
